big chunk renderer refactor
Not ready at all, but touches a lot and removes all the garbage junk around it. There are some todos left, but it is way cleaner (probably faster) and probably resolves various bugs

+++ src/main/java/de/bixilon/minosoft/gui/rendering/chunk/mesh/ChunkMesh.kt
@@ -17,7 +17,7 @@ import de.bixilon.minosoft.gui.rendering.system.base.buffer.vertex.VertexBuffer
 import de.bixilon.minosoft.gui.rendering.util.mesh.Mesh
 
 class ChunkMesh(buffer: VertexBuffer) : Mesh(buffer), Comparable<ChunkMesh> {
-    var distance: Float = 0.0f
+    var distance: Int = 0
 
     override fun compareTo(other: ChunkMesh): Int {
         if (distance < other.distance) return -1
